remove uneccesary "extract hypothesis answer" and simplify the class based arch
