+++ src/any_agent/evaluation/evaluate.py
@@ -3,11 +3,12 @@ from pydantic import BaseModel, ConfigDict
 from any_agent.config import AgentFramework
 from any_agent.evaluation.evaluation_case import EvaluationCase
 from any_agent.evaluation.evaluators import (
-    CheckpointEvaluator,
-    HypothesisEvaluator,
-    QuestionAnsweringSquadEvaluator,
+    evaluate_checkpoint,
+    evaluate_hypothesis,
+    evaluate_qa_squad,
 )
-from any_agent.evaluation.evaluators.schemas import EvaluationResult
+from any_agent.evaluation.schemas import EvaluationResult
+from any_agent.logging import logger
 from any_agent.tracing import TracingProcessor
 from any_agent.tracing.trace import AgentTrace
 
@@ -18,7 +19,7 @@ class TraceEvaluationResult(BaseModel):
     model_config = ConfigDict(extra="forbid")
 
     trace: AgentTrace
-    hypothesis_answer: str
+    hypothesis_answer: str | None
     checkpoint_results: list[EvaluationResult]
     hypothesis_answer_results: list[EvaluationResult]
     direct_results: list[EvaluationResult]
@@ -44,31 +45,39 @@ def evaluate(
     trace: AgentTrace,
     agent_framework: AgentFramework,
 ) -> TraceEvaluationResult:
-    checkpoint_evaluator = CheckpointEvaluator(model=evaluation_case.llm_judge)
-    hypothesis_evaluator = HypothesisEvaluator(model=evaluation_case.llm_judge)
-    qa_evaluator = QuestionAnsweringSquadEvaluator()
     processor = TracingProcessor.create(agent_framework)
     if not processor:
         msg = f"Processor for {agent_framework} not available."
         raise ValueError(msg)
-    hypothesis_answer = processor._extract_hypothesis_answer(trace=trace)
-    checkpoint_results = checkpoint_evaluator.evaluate(
+    hypothesis_answer = trace.final_output
+    checkpoint_results = evaluate_checkpoint(
+        model=evaluation_case.llm_judge,
         trace=trace,
         checkpoints=evaluation_case.checkpoints,
         processor=processor,
     )
-    hypothesis_answer_results = hypothesis_evaluator.evaluate(
-        hypothesis_final_output=hypothesis_answer,
-        ground_truth_answer_dict=evaluation_case.ground_truth,
-        ground_truth_checkpoints=evaluation_case.final_output_criteria,
-    )
+    if hypothesis_answer is not None:
+        hypothesis_answer_results = evaluate_hypothesis(
+            model=evaluation_case.llm_judge,
+            hypothesis_final_output=hypothesis_answer,
+            ground_truth_answer_dict=evaluation_case.ground_truth,
+            ground_truth_checkpoints=evaluation_case.final_output_criteria,
+        )
+    else:
+        logger.warning(
+            "No hypothesis answer found in the trace. Skipping hypothesis evaluation."
+        )
+        hypothesis_answer_results = []
 
-    if evaluation_case.ground_truth:
-        direct_results = qa_evaluator.evaluate(
+    if evaluation_case.ground_truth and hypothesis_answer is not None:
+        direct_results = evaluate_qa_squad(
             hypothesis_answer=hypothesis_answer,
             ground_truth_answer=evaluation_case.ground_truth,
         )
     else:
+        logger.warning(
+            "No ground truth answer found in the evaluation case or no hypothesis answer in the trace. Skipping direct evaluation."
+        )
         direct_results = []
 
     return TraceEvaluationResult(
