--- src/any_agent/evaluation/evaluators/__init__.py
@@ -1,9 +1,0 @@
-from .CheckpointEvaluator import CheckpointEvaluator
-from .HypothesisEvaluator import HypothesisEvaluator
-from .QuestionAnsweringSquadEvaluator import QuestionAnsweringSquadEvaluator
-
-__all__ = [
-    "CheckpointEvaluator",
-    "HypothesisEvaluator",
-    "QuestionAnsweringSquadEvaluator",
-]
