+++ tests/unit/evaluation/conftest.py
@@ -2,7 +2,8 @@ import json
 
 import pytest
 
-from any_agent.evaluation.evaluation_case import CheckpointCriteria, EvaluationCase
+from any_agent.evaluation.evaluation_case import EvaluationCase
+from any_agent.evaluation.schemas import CheckpointCriteria
 from any_agent.tracing.trace import AgentSpan, AgentTrace
 
 
@@ -26,4 +27,4 @@ def agent_trace() -> AgentTrace:
     with open(trace_path, encoding="utf-8") as f:
         spans = json.loads(f.read())
     spans = [AgentSpan.model_validate_json(span) for span in spans]
-    return AgentTrace(spans=spans)
+    return AgentTrace(spans=spans, final_output="Final output")
